fix: update C extension and improve test coverage

+++ ext/ext/klass/event/event_name.h
@@ -1,6 +1,8 @@
 #ifndef SFML_RB_EVENT_NAME_H
 #define SFML_RB_EVENT_NAME_H
 
-const char *get_event_name(unsigned long long int name);
+#include <stddef.h>
+
+const char *get_event_name(size_t name);
 
 #endif //SFML_RB_EVENT_NAME_H
